requirements.txt
fastapi
uvicorn
gunicorn
azure-data-tables
pydantic
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
